+++ python/ray/tune/config_parser.py
@@ -10,7 +10,8 @@ import os
 from six import string_types
 
 from ray.tune import TuneError
-from ray.tune.trial import Trial, json_to_resources
+from ray.tune.trial import Trial
+from ray.tune.resources import json_to_resources
 from ray.tune.logger import _SafeFallbackEncoder
